Создана папка utils для вспомогательных функций, рефакторинг
