+++ src/api/queries.ts
@@ -1,7 +1,6 @@
 import axios from "axios";
-import { format } from "date-fns";
 import { SERVER_URI } from "../constants/api-constants";
-import { ITodo } from "./mutations";
+import { formatTodosForAccordion } from "../utils/todos-utils";
 
 export const getTodos = async () => {
   const result = await axios.get(`${SERVER_URI}/todos`);
@@ -9,20 +8,4 @@ export const getTodos = async () => {
   const formatedTodos = formatTodosForAccordion(result.data);
 
   return formatedTodos;
-};
-
-const formatTodosForAccordion = (todos: ITodo[]) => {
-  let newTodos: { [key: string]: ITodo[] } = {};
-
-  todos.forEach((todo) => {
-    const key = format(new Date(todo.date), "MM/dd/yyyy");
-
-    if (!newTodos.hasOwnProperty(key)) {
-      newTodos[key] = [todo];
-    } else {
-      newTodos[key].push(todo);
-    }
-  });
-
-  return newTodos;
 };
